requirements.txt
# Django
Django>=4.2.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# Распознавание речи
openai-whisper>=20231117
torch>=2.0.0

# Обработка аудио
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0

# Работа с файлами
Pillow>=10.0.0

# Утилиты
python-decouple>=3.8

# Фоновые задачи
celery>=5.3.0
redis>=5.0.0

# PostgreSQL
psycopg2-binary>=2.9.0

# WSGI сервер для продакшена
gunicorn>=21.2.0

# Мониторинг и логирование
sentry-sdk>=1.32.0

# Кеширование для Django
django-redis>=5.2.0

